Concat operator is updated to support dynamic_shape 1) Added new concat_negative_indexing test 2) Replaced size_t -> int64_t 3) Support for a negative indexing, calculate axis = axis + int64(input_rank)    if (axis < 0) { axis = axis + int64_t(this_input_rank); }

+++ src/ngraph/op/concat.hpp
@@ -37,13 +37,13 @@ namespace ngraph
             ///
             /// \param args               The outputs producing the input tensors.
             /// \param axis The axis along which to concatenate the input tensors.
-            Concat(const OutputVector& args, size_t axis);
+            Concat(const OutputVector& args, int64_t axis);
 
             /// \brief Constructs a concatenation operation.
             ///
             /// \param args               The nodes producing the input tensors.
             /// \param axis The axis along which to concatenate the input tensors.
-            Concat(const NodeVector& args, size_t axis);
+            Concat(const NodeVector& args, int64_t axis);
 
             void validate_and_infer_types() override;
 
@@ -59,7 +59,7 @@ namespace ngraph
         protected:
             virtual void generate_adjoints(autodiff::Adjoints& adjoints,
                                            const NodeVector& deltas) override;
-            size_t m_axis;
+            int64_t m_axis;
         };
     }
 }
